+++ l10n_ro_hr_payroll/models/hr_contract.py
