+++ l10n_ro_hr_payroll/models/res_company.py
